requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-docx==1.2.0
python-multipart==0.0.9

# Faster-Whisper (lightweight Whisper reimplementation)
faster-whisper==1.0.3
